Refactor rendering to use separate functions for the variuos shapes
