+++ src/features/smartSearch/components/SmartSearchSankeyDiagram/index.tsx
@@ -1,6 +1,12 @@
 import { lighten } from '@mui/system';
 import { FC, useEffect, useRef } from 'react';
 
+import {
+  drawFilterInputOutput,
+  drawFilterMainStream,
+  Measurements,
+} from './drawing';
+
 type SmartSearchSankeyDiagramProps = {
   filterStats: {
     matched: number;
@@ -60,185 +66,37 @@ const SmartSearchSankeyDiagram: FC<SmartSearchSankeyDiagramProps> = ({
       const change = stats.output - inputCount;
       const changeWidth = (Math.abs(change) / maxSegOutput) * maxStreamWidth;
 
+      const measurements: Measurements = {
+        arrowDepth,
+        arrowWidth,
+        change,
+        changeWidth,
+        diagCenter,
+        diagWidth,
+        inputWidth,
+        lineWidth: 2,
+        margin,
+        offsetY: index * segHeight,
+        outputWidth,
+        segHeight,
+      };
+
       const gradient =
         index == mouseState.current.hoveredSegment
           ? hoverGradient
           : baseGradient;
-      context.fillStyle = gradient;
-      context.beginPath();
 
-      if (inputCount > 0 && change >= 0) {
-        // Main stream when adding
-        drawPath(
-          context,
-          [
-            ['M', diagCenter - inputWidth / 2, 0],
-            ['L', diagCenter + inputWidth / 2, 0],
-            [
-              'C',
-              [diagCenter + inputWidth / 2, 0.4 * segHeight],
-              [diagCenter + outputWidth / 2, 0.6 * segHeight],
-              [diagCenter + outputWidth / 2, segHeight],
-            ],
-            ['L', diagCenter - outputWidth / 2 + changeWidth, segHeight],
-            [
-              'C',
-              [diagCenter - outputWidth / 2 + changeWidth, 0.6 * segHeight],
-              [diagCenter - inputWidth / 2, 0.4 * segHeight],
-              [diagCenter - inputWidth / 2, 0],
-            ],
-          ],
-          index * segHeight
-        );
-      } else if (inputCount > 0 && change < 0) {
-        // Main stream when removing
-        drawPath(
-          context,
-          [
-            ['M', diagCenter - inputWidth / 2, 0],
-            ['L', diagCenter + inputWidth / 2 - changeWidth, 0],
-            [
-              'C',
-              [diagCenter + inputWidth / 2 - changeWidth, 0.4 * segHeight],
-              [diagCenter + outputWidth / 2, 0.6 * segHeight],
-              [diagCenter + outputWidth / 2, segHeight],
-            ],
-            ['L', diagCenter - outputWidth / 2, segHeight],
-            [
-              'C',
-              [diagCenter - outputWidth / 2, 0.6 * segHeight],
-              [diagCenter - inputWidth / 2, 0.4 * segHeight],
-              [diagCenter - inputWidth / 2, 0],
-            ],
-          ],
-          index * segHeight
-        );
-      }
-
-      if (change > 0) {
-        // Input when adding
-        drawPath(
-          context,
-          [
-            ['M', 0, segHeight / 2 + arrowWidth / 2],
-            ['L', arrowDepth, segHeight / 2],
-            ['L', 0, segHeight / 2 - arrowWidth / 2],
-            [
-              'C',
-              [
-                diagCenter - outputWidth / 2 + changeWidth / 2,
-                segHeight / 2 - arrowWidth / 2,
-              ],
-              [diagCenter - outputWidth / 2 + changeWidth, segHeight / 1.5],
-              [diagCenter - outputWidth / 2 + changeWidth, segHeight],
-            ],
-            ['L', diagCenter - outputWidth / 2, segHeight],
-            [
-              'C',
-              [diagCenter - outputWidth / 2, segHeight / 1.5],
-              [diagCenter - outputWidth / 2, segHeight / 2 + arrowWidth / 2],
-              [0, segHeight / 2 + arrowWidth / 2],
-            ],
-          ],
-          index * segHeight
-        );
-      } else if (change < 0) {
-        // Output when removing
-        drawPath(
-          context,
-          [
-            ['M', diagCenter + inputWidth / 2 - changeWidth, 0],
-            ['L', diagCenter + inputWidth / 2, 0],
-            [
-              'C',
-              [diagCenter + inputWidth / 2, segHeight / 3],
-              [diagWidth - margin, segHeight / 2 - arrowWidth / 2],
-              [diagWidth - arrowDepth, segHeight / 2 - arrowWidth / 2],
-            ],
-            ['L', diagWidth, segHeight / 2],
-            ['L', diagWidth - arrowDepth, segHeight / 2 + arrowWidth / 2],
-            [
-              'C',
-              [diagWidth - margin, segHeight / 2 + arrowWidth / 2],
-              [diagCenter + inputWidth / 2 - changeWidth, segHeight / 2],
-              [diagCenter + inputWidth / 2 - changeWidth, 0],
-            ],
-          ],
-          index * segHeight
-        );
-      }
-      context.fill();
-
-      if (change == 0) {
-        const lineWidth = 2;
+      if (inputCount > 0) {
+        context.fillStyle = gradient;
         context.beginPath();
-        context.strokeStyle = gradient;
-        context.lineWidth = lineWidth;
-        context.setLineDash([3, 3]);
-
-        // Change is zero, so we should render a placeholder
-        if (stats.op == 'add') {
-          drawPath(
-            context,
-            [
-              ['M', 0, segHeight / 2 + arrowWidth / 2],
-              ['L', arrowDepth, segHeight / 2],
-              ['L', 0, segHeight / 2 - arrowWidth / 2],
-              [
-                'C',
-                [
-                  diagCenter - outputWidth / 2 + arrowWidth / 2,
-                  segHeight / 2 - arrowWidth / 2,
-                ],
-                [diagCenter - outputWidth / 2 + arrowWidth, segHeight / 1.5],
-                [
-                  diagCenter - outputWidth / 2 + arrowWidth,
-                  segHeight - lineWidth / 2,
-                ],
-              ],
-              [
-                'L',
-                diagCenter - outputWidth / 2 + lineWidth / 2,
-                segHeight - lineWidth / 2,
-              ],
-              [
-                'C',
-                [diagCenter - outputWidth / 2, segHeight / 1.5],
-                [diagCenter - outputWidth / 2, segHeight / 2 + arrowWidth / 2],
-                [0, segHeight / 2 + arrowWidth / 2],
-              ],
-            ],
-            index * segHeight
-          );
-        } else if (stats.op == 'sub') {
-          drawPath(
-            context,
-            [
-              ['M', diagCenter + inputWidth / 2 - arrowWidth, lineWidth / 2],
-              ['L', diagCenter + inputWidth / 2 - lineWidth / 2, lineWidth / 2],
-              [
-                'C',
-                [diagCenter + inputWidth / 2, segHeight / 3],
-                [diagWidth - margin, segHeight / 2 - arrowWidth / 2],
-                [diagWidth - arrowDepth, segHeight / 2 - arrowWidth / 2],
-              ],
-              ['L', diagWidth, segHeight / 2],
-              ['L', diagWidth - arrowDepth, segHeight / 2 + arrowWidth / 2],
-              [
-                'C',
-                [diagWidth - margin, segHeight / 2 + arrowWidth / 2],
-                [diagCenter + inputWidth / 2 - arrowWidth, segHeight / 2],
-                [diagCenter + inputWidth / 2 - arrowWidth, 0],
-              ],
-            ],
-            index * segHeight
-          );
-        }
-
-        context.stroke();
+        drawFilterMainStream(context, measurements);
+        context.fill();
       }
+
+      drawFilterInputOutput(context, gradient, stats.op, measurements);
     });
 
+    // Draw the final arrow
     const lastStats = filterStats[filterStats.length - 1];
     const outputWidth = (lastStats.output / maxSegOutput) * maxStreamWidth;
     context.fillStyle = baseGradient;
@@ -295,34 +153,5 @@ const SmartSearchSankeyDiagram: FC<SmartSearchSankeyDiagramProps> = ({
     />
   );
 };
-
-type PathElem =
-  | ['C', [number, number], [number, number], [number, number]]
-  | ['M', number, number]
-  | ['L', number, number];
-
-function drawPath(
-  context: CanvasRenderingContext2D,
-  path: PathElem[],
-  offsetY = 0
-): void {
-  for (const elem of path) {
-    if (elem[0] == 'M') {
-      context.moveTo(elem[1], elem[2] + offsetY);
-    } else if (elem[0] == 'C') {
-      const [, c1, c2, p] = elem;
-      context.bezierCurveTo(
-        c1[0],
-        c1[1] + offsetY,
-        c2[0],
-        c2[1] + offsetY,
-        p[0],
-        p[1] + offsetY
-      );
-    } else if (elem[0] == 'L') {
-      context.lineTo(elem[1], elem[2] + offsetY);
-    }
-  }
-}
 
 export default SmartSearchSankeyDiagram;
